+++ resources/node.rb
@@ -19,17 +19,15 @@
 
 require 'ipaddress'
 
-actions :create
-default_action :create
-
 property :cluster_name, String, name_property: true
 property :cluster_ip, String, required: true, callbacks: { 'must be a valid IP address' => ->(ip) { IPAddress.valid?(ip) } }
-property :install_tools, [TrueClass, FalseClass], required: false, default: true
-property :quorum_disk, String, required: false
+property :install_tools, [TrueClass, FalseClass], default: true
+property :fs_witness, String
+property :quorum_disk, String
 property :run_as_user, String, required: true, default: lazy { node['windows_failover_cluster']['run_as_user'] }, callbacks: { 'must not be nil' => ->(p) { !p.nil? } }
 property :run_as_password, String, required: true, default: lazy { node['windows_failover_cluster']['run_as_password'] }, callbacks: { 'must not be nil' => ->(p) { !p.nil? } }
 
-actions :create, :join
+default_action :create
 
 action_class do
   def cluster_contain_node?
@@ -44,13 +42,28 @@ action_class do
     powershell_out_with_options('(Get-ClusterQuorum).QuorumResource.Name').stdout.chomp == disk_name
   end
 
-  def install_windows_feature(feature)
-    to_array(feature).each do |f|
-      windows_feature f do
-        install_method :windows_feature_powershell
-        action :install
-      end
-    end
+  def cluster_quorum_fs_witness?
+    powershell_out_with_options('(Get-ClusterResource "File Share Witness")').stdout.chomp =~ /File Share Witness/
+  end
+
+  def cluster_share_path?(share_path)
+    powershell_out_with_options('(Get-ClusterResource "File Share Witness" | Get-ClusterParameter SharePath).Value').stdout.chomp == share_path
+  end
+
+  def set_quorum_node_majority
+    powershell_out_with_options('Set-ClusterQuorum -NodeMajority')
+  end
+
+  def create_new_fs_witness
+    log "Configuring File Share Witness: #{new_resource.fs_witness}"
+    powershell_out_with_options!("Set-ClusterQuorum -NodeAndFileShareMajority \'#{new_resource.fs_witness}\'")
+  end
+
+  def install_windows_feature(features)
+    windows_feature [features].flatten do
+      install_method :windows_feature_powershell
+      action :nothing
+    end.run_action(:install)
   end
 
   def powershell_out_options
@@ -64,12 +77,6 @@ action_class do
   def powershell_out_with_options!(script)
     powershell_out!(script, powershell_out_options)
   end
-
-  def to_array(var)
-    var = var.is_a?(Array) ? var : [var]
-    var = var.reject(&:nil?)
-    var
-  end
 end
 
 action :create do
@@ -77,12 +84,31 @@ action :create do
   windows_features = %w(Failover-Clustering RSAT-Clustering-Powershell)
   windows_features << 'RSAT-Clustering-Mgmt' if new_resource.install_tools
   install_windows_feature(windows_features)
+
   # Create the cluster
   powershell_out_with_options!("New-Cluster -Name #{new_resource.cluster_name} -Node #{node['hostname']} -StaticAddress #{new_resource.cluster_ip} -Force") unless cluster_exist?(new_resource.cluster_name)
   # Add any available disks to the cluster
   powershell_out_with_options('Get-ClusterAvailableDisk | Add-ClusterDisk')
-  # Set quorum to use node & disk majority
-  powershell_out_with_options!("Set-ClusterQuorum -NodeAndDiskMajority \'#{new_resource.quorum_disk}\'") unless cluster_quorum_disk?(new_resource.quorum_disk)
+
+  if new_resource.quorum_disk && new_resource.fs_witness
+    Chef::Log.fatal('You provided both quorum_disk and fs_witness, only one is supported!')
+  end
+
+  # Configure quorum using node & disk majority
+  if new_resource.quorum_disk
+    powershell_out_with_options!("Set-ClusterQuorum -NodeAndDiskMajority \'#{new_resource.quorum_disk}\'") unless cluster_quorum_disk?(new_resource.quorum_disk)
+  end
+
+  # Configure quorum using node & file share majority
+  if new_resource.fs_witness
+    if cluster_quorum_fs_witness?
+      # Remove File Share Witness if it is not configured with the value we provided
+      log 'Resetting File Share Witness' unless cluster_share_path?(new_resource.fs_witness)
+      set_quorum_node_majority unless cluster_share_path?(new_resource.fs_witness)
+    end
+    # If we got here then a file share witness is not configured so we should configure it
+    create_new_fs_witness unless cluster_quorum_fs_witness?
+  end
 end
 
 action :join do
@@ -90,6 +116,7 @@ action :join do
   windows_features = %w(Failover-Clustering RSAT-Clustering-Powershell)
   windows_features << 'RSAT-Clustering-Mgmt' if new_resource.install_tools
   install_windows_feature(windows_features)
+
   # Join the cluster
   powershell_out_with_options!("Add-ClusterNode -Cluster #{new_resource.cluster_name} -Name #{node['hostname']}") if cluster_exist?(new_resource.cluster_name) && !cluster_contain_node?
 end