+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+- Add support to use a File Share Witness for quorum
+- Add support to configure multiple services for a single cluster resource
+
 ## 0.2.0 (2019-01-14)
 
 - Updated for Chef 14 and now requires a minimum version of Chef 14